+++ sample/ACPAudienceSample/android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.acpcoresample",
+    package = "com.acpaudiencesample",
 )
 
 android_resource(
     name = "res",
-    package = "com.acpcoresample",
+    package = "com.acpaudiencesample",
     res = "src/main/res",
 )
