+++ sample/ACPAudienceSample/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">ACPCoreSample</string>
+    <string name="app_name">ACPAudienceSample</string>
 </resources>
