+++ sample/ACPAudienceSample/android/settings.gradle
@@ -1,9 +1,7 @@
-rootProject.name = 'ACPCoreSample'
+rootProject.name = 'ACPAudienceSample'
 include ':@adobe_react-native-acpaudience'
 project(':@adobe_react-native-acpaudience').projectDir = new File(rootProject.projectDir, '../node_modules/@adobe/react-native-acpaudience/android')
 include ':@adobe_react-native-acpcore'
 project(':@adobe_react-native-acpcore').projectDir = new File(rootProject.projectDir, '../node_modules/@adobe/react-native-acpcore/android')
-include ':@adobe_react-native-acpanalytics'
-project(':@adobe_react-native-acpanalytics').projectDir = new File(rootProject.projectDir, '../node_modules/@adobe/react-native-acpanalytics/android')
 
 include ':app'
