+++ sample/ACPAudienceSample/ios/ACPAudienceSample/Images.xcassets/AppIcon.appiconset/Contents.json
