+++ sample/ACPAudienceSample/ios/ACPAudienceSample/Images.xcassets/Contents.json
