+++ sample/ACPAudienceSample/ios/ACPAudienceSampleTests/Info.plist
